requirements.txt
PyQt6
numpy
matplotlib
librosa
Pillow
opencv-python
